+++ tests/JavascriptGeneratorBundle/Fixtures/expected/named-scalar/src/service/NamedScalarClient.js
@@ -1,0 +1,34 @@
+import { createRequest, ClientWrapper } from '@paysera/http-client-common';
+
+import Payment from '../entity/Payment';
+
+class NamedScalarClient {
+    /**
+     * @param {ClientWrapper} client
+     */
+    constructor(client) {
+        this.client = client;
+    }
+
+    /**
+     * Create payment
+     * POST /payments
+     *
+     * @param {Payment} payment
+     * @return {Promise.<Payment>}
+     */
+    createPayment(payment) {
+        const request = createRequest(
+            'POST',
+            `payments`,
+            payment,
+        );
+
+        return this.client
+            .performRequest(request)
+            .then(data => new Payment(data));
+    }
+
+}
+
+export default NamedScalarClient;