+++ tests/JavascriptGeneratorBundle/Fixtures/expected/named-scalar/src/service/createClient.js
@@ -1,0 +1,36 @@
+import { createClient } from '@paysera/http-client-common';
+import NamedScalarClient from './NamedScalarClient';
+
+/**
+ * @param {string} baseURL
+ * @param {[]|null} middleware
+ * @param {object} options
+ *
+ * @returns {NamedScalarClient}
+ */
+/* eslint import/prefer-default-export: ["off"] */
+export const createNamedScalarClient = ({
+    baseURL = 'https://my-api.example.com/rest/v1/',
+    middleware = null,
+    options = {}
+}) => {
+    const defaultUrlParameters = {};
+    
+    if (Object.prototype.hasOwnProperty.call(options, 'urlParameters')) {
+        const { urlParameters } = options;
+        for (let [key, value] of Object.entries(defaultUrlParameters)) {
+            if (!Object.prototype.hasOwnProperty.call(urlParameters, key)) {
+                urlParameters[key] = value;
+            }
+        }
+        options.urlParameters = urlParameters;
+    }
+
+    return new NamedScalarClient(
+        createClient({
+            baseURL,
+            middleware,
+            options
+        })
+    )
+};